New functions and a split (cont'd)
Migrate some functionality to a new Statik class, for shared Statik
handling across any implementation. This class is a lovely abstract class
for all your implementation needs.

Create the actual Stat class, for stat tracking, which is neat.
Additionally, a bunch of basic classes have been created for this, though
the Stat class can also just be subclassed as it is a fancy abstract class
just waiting for some attention.

Added processing of data, getting the system most of the way there.

Lastly, added 50% more kittens.